+++ source/isaaclab_contrib/isaaclab_contrib/deformable/cloner_hooks.py
