+++ source/isaaclab_contrib/isaaclab_contrib/deformable/model_cfg_hook.py
