+++ source/isaaclab_contrib/isaaclab_contrib/deformable/newton_manager_cfg.py
