+++ source/isaaclab_newton/isaaclab_newton/assets/deformable_object/__init__.py
@@ -5,12 +5,12 @@
 
 """Newton deformable object backend.
 
-The implementation lives in :mod:`isaaclab_experimental.deformable`.
+The implementation lives in :mod:`isaaclab_contrib.deformable`.
 This stub re-exports the class so the :class:`FactoryBase` auto-discovery
 (which tries ``isaaclab_newton.assets.deformable_object``) finds it.
 """
 
-from isaaclab_experimental.deformable.deformable_object import DeformableObject
-from isaaclab_experimental.deformable.deformable_object_data import DeformableObjectData
+from isaaclab_contrib.deformable.deformable_object import DeformableObject
+from isaaclab_contrib.deformable.deformable_object_data import DeformableObjectData
 
 __all__ = ["DeformableObject", "DeformableObjectData"]
