+++ source/isaaclab_tasks_experimental/isaaclab_tasks_experimental/direct/pick_cloth/pick_cloth_env.py
@@ -13,7 +13,7 @@ from collections.abc import Sequence
 import torch
 import warp as wp
 
-from isaaclab_experimental.deformable import register_hooks as _register_deformable_hooks
+from isaaclab_contrib.deformable import register_hooks as _register_deformable_hooks
 
 _register_deformable_hooks()
 
@@ -41,7 +41,7 @@ class PickClothEnv(DirectRLEnv):
         # Without a robot the coupled solver (rigid + VBD) is unnecessary and will
         # fail because there are no rigid bodies.  Swap to VBD-only automatically.
         if not self._has_robot:
-            from isaaclab_experimental.deformable.newton_manager_cfg import CoupledSolverCfg
+            from isaaclab_contrib.deformable.newton_manager_cfg import CoupledSolverCfg
 
             physics_cfg = cfg.sim.physics
             if hasattr(physics_cfg, "solver_cfg") and isinstance(physics_cfg.solver_cfg, CoupledSolverCfg):
