+++ source/isaaclab_tasks_experimental/isaaclab_tasks_experimental/direct/pick_cloth/pick_cloth_env_cfg.py
@@ -8,7 +8,7 @@
 import importlib.util
 import os.path
 
-from isaaclab_experimental.deformable.newton_manager_cfg import CoupledSolverCfg, NewtonModelCfg, VBDSolverCfg
+from isaaclab_contrib.deformable.newton_manager_cfg import CoupledSolverCfg, NewtonModelCfg, VBDSolverCfg
 from isaaclab_newton.physics import FeatherstoneSolverCfg, MJWarpSolverCfg, NewtonCfg
 from isaaclab_visualizers.newton import NewtonVisualizerCfg
